requirements.txt
fastapi
uvicorn
transformers
Pillow
torch
